Initial commit.
It builds and DKMSes just fine, but my kernel here really doesn't
like these modules for some reason, it refuses to find the new
devices and bind them to the right drivers accordingly.

I can rebind manually and things work as expected, but the kernel
doesn't seem to update the device matching.